--- Classic_RPG/src/game/ManaPool.java
@@ -1,23 +1,0 @@
-package game;
-
-public class ManaPool{
-	private int manaMax;
-	private int mana;
-	
-	protected ManaPool(int size) {
-		this.mana = size;
-		this.manaMax = size;
-	}
-
-	public int getMana() {
-		return mana;
-	}
-
-	public void setMana(int mana) {
-		this.mana = mana;
-	}
-
-	public int getManaMax() {
-		return manaMax;
-	}
-}
