+++ Classic_RPG/src/game/Wizard.java
@@ -11,14 +11,12 @@ public class Wizard extends Player{
 	public void specialHit(Entity target) {
 		attack(target, characteristics.getIntelligence() * 3);
 		
-		int currentMana = manaPool.getMana();
-		if(currentMana  - 75 < 0) {
-			manaPool.setMana(0);
+		if(getMana() - 75 < 0) {
+			setMana(0);
 		}
 		else {
-			manaPool.setMana(currentMana - 75);
+			setMana(getMana() - 75);
 		}
-		
 	}
 	
 	public void receiveAttack(int dmgIncoming) {
@@ -38,7 +36,7 @@ public class Wizard extends Player{
 	public void setDefensePosition(boolean defense_position_){
 		if(defense_position_) {
 			magicSheildHp = characteristics.getIntelligence() * 2;
-			manaPool.setMana(manaPool.getMana() - 55);
+			setMana(getMana() - 55);
 		}
 		super.setDefensePosition(defense_position_);
 	}
